support loading the eta database as stream
This allows for much better performance when we do not need
to load everything at once in order to read the database.

Also modified etadb-load to be just a chain of stream->list and
etadb-input-port->session-stream. For some reason this also seems
to perform better even when we read everything.